+++ src/app/page.tsx
@@ -1,6 +1,16 @@
 import { compareDesc } from 'date-fns';
 import { allPosts } from 'contentlayer/generated';
 import PostCard from '@/components/PostCard';
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  openGraph: {
+    title: 'SON의 개발 블로그',
+    description: '개발하면서 느낀점, 배운점을 기록합니다.',
+    images: 'https://source.unsplash.com/random/300×300',
+  },
+};
+
 export default function Home() {
   const posts = allPosts.sort((a, b) =>
     compareDesc(new Date(a.date), new Date(b.date)),
