+++ src/app/projects/[slug]/page.tsx
@@ -1,4 +1,3 @@
-import HeadOG from '@/components/HeadOG';
 import LinkCopy from '@/components/LinkCopy';
 import StakList from '@/components/StakList';
 import WritingList from '@/components/WritingList';
@@ -14,14 +13,21 @@ type Props = {
 
 export const generateMetadata = ({ params }: Props) => {
   const oneProject = projectObj.find(item => item.link === params.slug);
-  return { title: oneProject!.name, description: oneProject!.description };
+  return {
+    title: oneProject!.name,
+    description: oneProject!.description,
+    openGraph: {
+      title: oneProject?.name,
+      images: 'https://source.unsplash.com/random/300×300',
+      description: oneProject?.description,
+    },
+  };
 };
 
 export default function ProjectDetailPages({ params }: Props) {
   const oneProject = projectObj.find(item => item.link === params.slug);
   return (
     <article className="py-4">
-      <HeadOG title={oneProject!.name} description={oneProject?.description} />
       <div className="flex relyativy gap-2 items-center">
         <h1 className="font-black mb-5 text-3xl">{oneProject!.name}</h1>
         <LinkCopy />
